refactor(api): move url-builder
to src/shared
