+++ api/lib/domain/usecases/create-and-reconcile-user-to-organization-learner.js
@@ -3,7 +3,7 @@ const { isNil } = lodash;
 
 import { User } from '../../../src/identity-access-management/domain/models/User.js';
 import { EntityValidationError } from '../../../src/shared/domain/errors.js';
-import { getCampaignUrl } from '../../infrastructure/utils/url-builder.js';
+import { getCampaignUrl } from '../../../src/shared/infrastructure/utils/url-builder.js';
 import { STUDENT_RECONCILIATION_ERRORS } from '../constants.js';
 import {
   AlreadyRegisteredEmailError,
